Move AsyncReadCompletion to internal utils, trim comments
spi/ is the contract surface providers implement against; the whenComplete
error-safety helper is internal plumbing, so it belongs in datasources.utils
alongside ContentRangeParser (already shared cross-module). No behaviour
change. Also trims the helper's javadoc and the transformer's Synchronization
note down to the house style.
