+++ x-pack/plugin/esql-datasource-s3/src/main/java/org/elasticsearch/xpack/esql/datasource/s3/S3StorageObject.java
@@ -34,7 +34,7 @@ import java.io.InputStream;
 import java.time.Instant;
 import java.util.concurrent.Executor;
 
-import static org.elasticsearch.xpack.esql.datasources.spi.AsyncReadCompletion.errorSafe;
+import static org.elasticsearch.xpack.esql.datasources.utils.AsyncReadCompletion.errorSafe;
 
 /**
  * StorageObject implementation for S3 using AWS SDK v2.
